backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
python-dotenv>=1.0
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
dj-database-url
